Add files via upload
in this file, we are going to implement Andrew NG machine learning course example in python.
week 1,

Gradient descent is an important parameter in machine learning. it is used to minimize functions and update important parameters. The goal of gradient descent is to take the steepest possible way to achieve convergence. Here we will implement how gradient descent minimizes the cost function. 
in this example we want to predict Y using Gradient descent